requirements.txt
flask
python-dotenv
langchain-groq
